backend/requirements.txt
fastapi
uvicorn
google-auth
openai
python-dotenv
sqlalchemy
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
requests
requests-oauthlib
itsdangerous
chromadb
sentence-transformers
google-generativeai
